requirements.txt
# Image Description Bot - Requirements
# Python >=3.10,<3.12

# Web Framework
flask==3.1.2
flask-cors==6.0.1
gunicorn==23.0.0
werkzeug==3.1.4

# Environment & Configuration
python-dotenv==1.2.1

# Computer Vision & Face Detection
opencv-python==4.8.1
deepface==0.0.79
mediapipe==0.10.14
retina-face==0.0.17
mtcnn==1.0.0

# Deep Learning Frameworks
torch==2.7.1
tensorflow==2.15.0
tensorflow-intel==2.15.0

# Transformers & LLM
transformers==4.57.3
accelerate==1.12.0
bitsandbytes==0.49.1
sentencepiece==0.2.1
safetensors==0.7.0
tokenizers==0.22.1
huggingface-hub==0.36.0

# Data Processing
numpy==1.26.4
pandas==2.3.3
pillow==10.2.0

# Utilities
tqdm==4.67.1
pyyaml==6.0.3
requests==2.32.5

# Development Dependencies (optional - uncomment if needed)
# pytest==9.0.2
# pytest-cov==7.0.0
# pytest-flask==1.3.0
# pytest-mock==3.15.1
# coverage==7.12.0
# black==25.12.0
# ruff==0.14.8
# mypy==1.19.0